requirements.txt
spotipy==2.4.4
ipdb==0.10.3
python-dateutil==2.6.0
google-api-python-client==1.6.4
